+++ CHANGELOG.md
@@ -31,7 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Playlist and search downloads (Spankbang)
+- No videos detected on profiles, playlists or search results (Spankbang)
 
 ## [10.0.2] - 2026-07-05